restructured tests for einstein
