+++ test/contrib/einstein/test_einstein_kernels.py
@@ -2,6 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from collections import namedtuple
+from copy import copy
 
 import numpy as np
 from numpy.testing import assert_allclose
@@ -9,6 +10,7 @@ import pytest
 
 from jax import numpy as jnp, random
 
+from numpyro.contrib.einstein import SteinVI
 from numpyro.contrib.einstein.kernels import (
     GraphicalKernel,
     HessianPrecondMatrix,
@@ -19,6 +21,8 @@ from numpyro.contrib.einstein.kernels import (
     RandomFeatureKernel,
     RBFKernel,
 )
+from numpyro.infer import Trace_ELBO
+from numpyro.optim import Adam
 
 T = namedtuple("TestSteinKernel", ["kernel", "particle_info", "loss_fn", "kval"])
 
@@ -94,3 +98,26 @@ def test_kernel_forward(
     kernel_fn = kernel.compute(particles, particle_info(d), loss_fn)
     value = kernel_fn(*tparticles)
     assert_allclose(value, jnp.array(kval[mode]), rtol=1e-6)
+
+
+@pytest.mark.parametrize(
+    "kernel, particle_info, loss_fn, kval", TEST_CASES, ids=TEST_IDS
+)
+@pytest.mark.parametrize("mode", ["norm", "vector", "matrix"])
+@pytest.mark.parametrize("particles, tparticles", PARTICLES)
+def test_apply_kernel(
+    kernel, particles, particle_info, loss_fn, tparticles, mode, kval
+):
+    if mode not in kval:
+        pytest.skip()
+    (d,) = tparticles[0].shape
+    kernel_fn = kernel(mode=mode)
+    kernel_fn.init(random.PRNGKey(0), particles.shape)
+    kernel_fn = kernel_fn.compute(particles, particle_info(d), loss_fn)
+    v = np.ones_like(kval[mode])
+    stein = SteinVI(id, id, Adam(1.0), Trace_ELBO(), kernel(mode))
+    value = stein._apply_kernel(kernel_fn, *tparticles, v)
+    kval_ = copy(kval)
+    if mode == "matrix":
+        kval_[mode] = np.dot(kval_[mode], v)
+    assert_allclose(value, kval_[mode], rtol=2e-6)
