+++ test/contrib/einstein/test_einstein_util.py
@@ -46,9 +46,26 @@ def test_posdef(m):
     assert jnp.alltrue(jnp.linalg.eigvals(pd_m) > 0)
 
 
+@pytest.mark.parametrize("batch_shape", [(), (2,), (3, 1)])
+def test_posdef_shape(batch_shape):
+    dim = 4
+    x = np.random.normal(size=batch_shape + (dim, dim + 1))
+    m = x @ np.swapaxes(x, -2, -1)
+    assert_allclose(posdef(m), m, rtol=1e-5)
+
+
 @pytest.mark.parametrize("m", matrices)
 def test_sqrth(m):
     assert_allclose(sqrth(m), scipy.linalg.sqrtm(posdef(m)), atol=1e-5)
+
+
+@pytest.mark.parametrize("batch_shape", [(), (2,), (3, 1)])
+def test_sqrth_shape(batch_shape):
+    dim = 4
+    x = np.random.normal(size=batch_shape + (dim, dim + 1))
+    m = x @ np.swapaxes(x, -2, -1)
+    s = sqrth(m)
+    assert_allclose(s @ np.swapaxes(s, -2, -1), m, rtol=1e-5)
 
 
 @pytest.mark.parametrize("m", pd_matrices)
@@ -57,6 +74,17 @@ def test_sqrt_inv_sqrth(m):
     assert_allclose(msqrt, scipy.linalg.sqrtm(m), atol=1e-5)
     assert_allclose(minv, np.linalg.inv(m), atol=1e-4)
     assert_allclose(minv_sqrt, np.linalg.inv(scipy.linalg.sqrtm(m)), atol=1e-5)
+
+
+@pytest.mark.parametrize("batch_shape", [(), (2,), (3, 1)])
+def test_sqrth_and_inv_sqrth_shape(batch_shape):
+    dim = 4
+    x = np.random.normal(size=batch_shape + (dim, dim + 1))
+    m = x @ np.swapaxes(x, -2, -1)
+    s, i, si = sqrth_and_inv_sqrth(m)
+    assert_allclose(s @ np.swapaxes(s, -2, -1), m, rtol=1e-5)
+    assert_allclose(i, np.linalg.inv(m), rtol=1e-5)
+    assert_allclose(si @ np.swapaxes(si, -2, -1), i, rtol=1e-5)
 
 
 @pytest.mark.parametrize(
